IOS: FIX for https://javafx-jira.kenai.com/browse/RT-34010

+++ modules/web/src/ios/java/javafx/scene/web/JS2JavaBridge.java
@@ -117,7 +117,7 @@ class JS2JavaBridge {
         sb = sb.append(getJavaBridge()).append(".exportJSObject(").append(
              getJavaBridge()).append("['").append(jsName).append("'])");
         
-        Long jsId = (Long) webEngine.executeScript(sb.toString());
+        Integer jsId = (Integer) webEngine.executeScript(sb.toString());
         exportedObjectsByJSIds.put(jsId.toString(), jsObj);
         jsIdsByExportedObjects.put(jsObj, jsId.toString());
         log("populateObject<<executeScript");
